Tests: one shared force-until-moved ctime helper (tests/fs_clock.py) for the thread mail-off test and the two memo tests that carried identical copies
The thread mail-off test's chat-identity case slept 20 ms and hoped the
filesystem clock had ticked before its chmod (the review's low on the fifth
follow-up); the repo already had the bounded helper twice, in the timeline lane
memo test and the cleared-set memo test. It is lifted to tests/fs_clock.py:
flip the mode until ctime differs, bounded at 5 s, failing loudly on a
filesystem that never ticks ctime under chmod. The three tests import it.
